String splitting, New word, keypressed
Things i did:
I added a print for the typed characters below the currentcharacters.
i added the variable characterIndex which i use in keypressed to go through the elements of currentcharacters.
in they keypressed function i check if key=currentcharacter[characterindex] and if so then i add they key to typedcharacters.
I also added the function splitstring to split a string in the function setNewWord